Rework time

+++ src/timer.hpp
@@ -3,12 +3,12 @@
 namespace Gts {
 	class Timer {
 		public:
-			Timer(float delta);
+			Timer(double delta);
 			bool ShouldRun();
-			float TimeDelta();
+			double TimeDelta();
 		private:
-			float last_time = 0.0;
-			float delta = 0.01666;
-			float elaped_time = 0.0;
+			double last_time = 0.0;
+			double delta = 0.01666;
+			double elaped_time = 0.0;
 	};
 }
